requirements.txt
flask
flask-restful
GitPython>=0.3.1